# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AITeachMe",
    "description": "本地优先的 AI 助教后端服务。",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "注册",
        "description": "基于 device_key 的匿名身份升级为邮箱账号。",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSessionData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "当前资源状态不允许执行该操作。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "登录",
        "description": "邮箱密码登录，并绑定当前 device_key。",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSessionData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "接口可能返回业务错误。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "登出",
        "description": "清除登录态后回到 device_key 匿名身份。",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSessionData_"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/user": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "当前用户",
        "description": "读取当前 token/device_key 对应的用户会话信息。",
        "operationId": "user_api_v1_auth_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSessionData_"
                }
              }
            }
          },
          "401": {
            "description": "接口可能返回业务错误。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_AuthSessionData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthSessionData"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[AuthSessionData]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "AuthSessionData": {
        "properties": {
          "auth_enabled": {
            "type": "boolean",
            "title": "Auth Enabled",
            "description": "是否启用鉴权。"
          },
          "auth_ready": {
            "type": "boolean",
            "title": "Auth Ready",
            "description": "鉴权是否就绪。"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "令牌类型。",
            "default": "bearer"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token",
            "description": "访问令牌。"
          },
          "current_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeUser"
              },
              {
                "type": "null"
              }
            ],
            "description": "当前用户。"
          }
        },
        "type": "object",
        "required": [
          "auth_enabled",
          "auth_ready"
        ],
        "title": "AuthSessionData",
        "description": "鉴权会话数据。"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ErrorResponse",
        "description": "统一错误响应结构。",
        "example": {
          "code": 404,
          "message": "资源不存在"
        }
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "邮箱。"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password",
            "description": "密码。"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "登录请求。",
        "example": {
          "email": "user@example.com",
          "password": "change-me"
        }
      },
      "LogoutRequest": {
        "properties": {},
        "type": "object",
        "title": "LogoutRequest",
        "description": "登出请求。",
        "example": {}
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "邮箱。"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password",
            "description": "密码。"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "注册请求。",
        "example": {
          "email": "user@example.com",
          "password": "change-me"
        }
      },
...
      "RuntimeUser": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "用户 ID。"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "邮箱。"
          },
          "is_local": {
            "type": "boolean",
            "title": "Is Local",
            "description": "是否为本地模式用户。"
          },
          "device_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Key",
            "description": "设备标识。"
          },
          "is_authenticated": {
            "type": "boolean",
            "title": "Is Authenticated",
            "description": "是否已完成账号登录。",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "is_local"
        ],
        "title": "RuntimeUser",
        "description": "当前运行时用户。"
      },
...
    }
  }
}